Make "disable submit button" more specific to freehanddrawing plugin
Disabling the submit button on various Moodle forms, causes the button to become hidden. which is not the intended behavior we are looking for in those cases.

It seems to originate in this styles.css file, after installing question type freehanddrawing on our systems.

I am suggesting the above fix, although I am not sure it is the correct fix since I was not sure which submit button you intended to hide. So please see that I am using the correct page id.